+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.8
+
+- **FIXED**: Invalidate and relayout render object after system fonts changed.
+
 ## 0.0.7
 
 - **FIXED**: Preserved indentation on line breaks within list items [#4].
@@ -7,7 +11,7 @@
 
 ## 0.0.6
 
-- **FIXED**: Fixed escaping of special characters. [#6] 
+- **FIXED**: Fixed escaping of special characters. [#6]
 
 ## 0.0.5